+++ main.py
@@ -8,8 +8,6 @@ Eric Fernandes Evaristo (ErFer7)
 Luis Henrique Goulart Stemmer (lust2k)
 '''
 
-VERSION = "v0.4"
-
 import gi
 
 gi.require_version("Gtk", "3.0")